// STEAMZ-UG/src/types.ts
export interface MenuItemOptionChoice {
  id?: string;
  name: string;
  price: number;
  isDefault?: boolean;
  calories?: number;
}

export interface MenuItemOption {
  id?: string;
  name: string;
  required?: boolean;
  multiSelect?: boolean;
  maxSelect?: number;
  minSelect?: number;
  choices: MenuItemOptionChoice[];
}